clang tidy

+++ tasks/smyshlaev_a_sle_cg_seq/seq/src/ops_seq.cpp
@@ -73,7 +73,7 @@ bool SmyshlaevASleCgTaskSEQ::ValidationImpl() {
 bool SmyshlaevASleCgTaskSEQ::PreProcessingImpl() {
   const auto &a = GetInput().A;
   size_t n = a.size();
-  flat_A_.resize(static_cast<size_t>(n * n));
+  flat_A_.resize(n * n);
   for (size_t i = 0; i < n; ++i) {
     for (size_t j = 0; j < n; ++j) {
       flat_A_[(i * n) + j] = a[i][j];
